soft delete

+++ app/Http/Controllers/PostsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use App\Models\Post;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -16,6 +17,15 @@ class PostsController extends Controller
 
     public function getPostsByUser(Request $request, $userId) {
         $posts = Post::where('user_id', $userId)->with(['user', 'categories'])->paginate(8);
+        return view('posts', ['posts' => $posts]);
+    }
+
+    public function getPostsByCategory(Request $request, $categoryId) {
+
+        $posts = Post::whereHas('categories', function (Builder $query) use($categoryId) {
+            $query->where('category_id', $categoryId);
+        })->paginate(8);
+
         return view('posts', ['posts' => $posts]);
     }
 
@@ -63,23 +73,42 @@ class PostsController extends Controller
     }
 
     public function editPost(Request $request, $id) {
+
+        $request->validate([
+            'title' => 'required|string',
+            'body' => 'required|string',
+            'categories' => 'sometimes|array'
+        ]);
+
         $post = Post::with(['user', 'categories'])->findOrFail($id);
 
         $title = $request->title;
         $body = $request->body;
+        $categoryIds = $request->categories ? $request->categories : [];
 
         $post->title = $title;
         $post->body = $body;
+        $post->categories()->sync($categoryIds);
+
         $post->save();
 
-        return view('editPost', ['post' => $post, 'message' => 'Post successfully edited!']);
+        $categories = Category::all();
+
+        return view('editPost', ['post' => $post->refresh(), 'message' => 'Post successfully edited!', 'categories' => $categories]);
     }
 
     public function deletePost(Request $request, $id) {
-        $post = Post::findOrFail($id);
+        $post = Post::withoutGlobalScopes()->findOrFail($id);
+        $message = 'Post successfully soft deleted!';
 
-        $post->delete();
+        if ($post->deleted) {
+            $post->delete();
+            $message = 'Post successfully permanently deleted!';
+        } else {
+            $post->deleted = 1;
+            $post->save();
+        }
 
-        return redirect()->back()->with('message', 'Post successfully deleted!');
+        return redirect()->back()->with('message', $message);
     }
 }
